refactor: modify directory location
